create module for registering command under new architecture
